feat: add Link to bookItem, create bookItem component
